backend/requirements.txt
# Core backend framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
websockets==12.0
python-dotenv==1.0.0

# Vision models and processing
ultralytics==8.0.227
clip @ git+https://github.com/openai/CLIP.git
paddleocr==2.7.0.3
paddlepaddle==2.5.1  # CPU-only PaddlePaddle (use this if you don't have a GPU)

# Deep learning and transformers
transformers==4.36.0
accelerate==0.25.0
bitsandbytes==0.41.2.post2

# Data processing
pillow==10.1.0
# pinned to a version compatible with both ultralytics and paddleocr
# using headless to avoid GUI deps on servers/CI
opencv-python-headless==4.6.0.66
numpy==1.24.3
scipy==1.11.4
scikit-image==0.22.0

# Monitoring and utilities
GPUtil==1.4.0
psutil==5.9.6
tqdm==4.66.1

# Development tools
jupyter==1.0.0
ipython==8.18.1
pytest==7.4.3
pytest-asyncio==0.21.1

# NOTES & NEXT STEPS:
# 1) This file is configured for a NO-GPU (CPU-only) environment:
#    - paddlepaddle==2.5.1 (CPU) replaces paddlepaddle-gpu
#    - opencv-python-headless==4.6.0.66 satisfies both ultralytics and paddleocr and
#      avoids GUI-related dependencies. If you need GUI support locally, replace with
#      opencv-python==4.6.0.66.
# 2) Save this file as backend/requirements.txt (overwrite the old one).
# 3) Recommended install workflow (you're already in a virtualenv named ENV):
#      python -m pip install --upgrade pip setuptools wheel
#      pip install -r backend/requirements.txt --upgrade --force-reinstall
#    If pip's resolver still struggles, install the big packages first:
#      pip install paddlepaddle==2.5.1
#      pip install ultralytics==8.0.227
#      pip install paddleocr==2.7.0.3
#      pip install -r backend/requirements.txt
# 4) Windows-specific note for PaddlePaddle: if the pip install of paddlepaddle fails,
#    try using an official mirror (for example, Baidu mirror):
#      pip install paddlepaddle==2.5.1 -i https://mirror.baidu.com/pypi/simple
# 5) Quick verification after install:
#      python -c "import paddle; print('paddle', paddle.__version__)"
#      python -c "import cv2; print('cv2', cv2.__version__)"
#      python -c "import ultralytics; print('ultralytics', ultralytics.__version__)"
# 6) Run your FastAPI app locally with uvicorn:
#      uvicorn backend.main:app --reload --host 0.0.0.0 --port 8000
# 7) If you need GPU later, you can switch back to paddlepaddle-gpu and the
#    appropriate CUDA toolkit-compatible wheel.

# If you want, I can also:
#  - switch back to non-headless opencv (for local GUI needs), or
#  - produce a one-liner to install the most finicky packages first.
# Tell me which and I will update this file again.
opencv-python>=4.8.0
python-multipart>=0.0.6
